requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
drf-spectacular>=0.28
django-filter>=23.1
python-slugify>=8.0